backend/requirements.txt
# requirements.txt
# Legalyze API Dependencies

# ══════════════════════════════════════════════════════════════════
# Core Framework
# ══════════════════════════════════════════════════════════════════
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0

# ══════════════════════════════════════════════════════════════════
# Database
# ══════════════════════════════════════════════════════════════════
motor==3.3.2
pymongo==4.6.1

# ══════════════════════════════════════════════════════════════════
# Authentication & Security
# ══════════════════════════════════════════════════════════════════
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
cryptography==41.0.7

# ══════════════════════════════════════════════════════════════════
# File Processing
# ══════════════════════════════════════════════════════════════════
# PDF
PyMuPDF==1.23.8
pypdf==3.17.4

# DOCX
python-docx==1.1.0

# OCR
pytesseract==0.3.10
Pillow==10.2.0

# Document Generation
reportlab==4.0.9

# Excel (optional)
openpyxl==3.1.2

# ══════════════════════════════════════════════════════════════════
# AI & NLP
# ══════════════════════════════════════════════════════════════════
# Transformers
transformers==4.36.2
torch==2.1.2
tokenizers==0.15.0

# Sentence Transformers
sentence-transformers==2.3.1

# spaCy
spacy==3.7.2

# Vector Store
faiss-cpu==1.7.4
# Use faiss-gpu==1.7.4 if you have CUDA

# ══════════════════════════════════════════════════════════════════
# Cloud Storage
# ══════════════════════════════════════════════════════════════════
boto3==1.34.26
botocore==1.34.26

# Firebase (optional)
firebase-admin==6.3.0

# ══════════════════════════════════════════════════════════════════
# Middleware & Utilities
# ══════════════════════════════════════════════════════════════════
# Rate Limiting
slowapi==0.1.9

# Email
aiosmtplib==3.0.1

# Environment Variables
python-dotenv==1.0.0

# HTTP
httpx==0.26.0
aiohttp==3.9.1

# Async utilities
asyncio==3.4.3

# ══════════════════════════════════════════════════════════════════
# Data Processing
# ══════════════════════════════════════════════════════════════════
numpy==1.26.3
pandas==2.1.4
scikit-learn==1.4.0

# ══════════════════════════════════════════════════════════════════
# Development & Testing (optional)
# ══════════════════════════════════════════════════════════════════
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0
black==23.12.1
flake8==7.0.0
mypy==1.8.0

# ══════════════════════════════════════════════════════════════════
# Monitoring (optional)
# ══════════════════════════════════════════════════════════════════
sentry-sdk[fastapi]==1.39.2
prometheus-fastapi-instrumentator==6.1.0